crm-clean/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
langchain==0.2.6
langchain-groq==0.1.6
langchain-core==0.2.10
langgraph==0.1.19
httpx==0.27.0
python-dotenv==1.0.1
